Tests: Move "utils" tests into dedicated folder
